requirements.txt
attrs==19.1.0
certifi==2019.6.16
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==2.2.3
django-filter==2.2.0
djangorestframework==3.10.0
drf-flex-fields==0.5.0
drf-yasg==1.16.1
factory-boy==2.12.0
Faker==2.0.0
idna==2.8
inflection==0.3.1
itypes==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
packaging==19.1
pyparsing==2.4.2
python-dateutil==2.8.0
pytz==2019.2
requests==2.22.0
ruamel.yaml==0.16.0
ruamel.yaml.clib==0.1.0
six==1.12.0
sqlparse==0.3.0
text-unidecode==1.2
uritemplate==3.0.0
urllib3==1.25.3
